Added LineDetector.java, SetalyzerImagecapture.java, SetalyzerShutterCallback.java
SetalyzerImageCapture and ShutterCallback are classes that can be used
as camera callbacks (they implement ShutterCallback and
PictureCallback).

Updated SetViewerActivity to have the dummy button take a picture using
Setalyzer... as callbacks.

LineDetector is experimentation with boofcv line detection and a vague
attempt to display an image back (doesn't work)

Also added an ImageView to activity_set_viewer, which should be broken
out into a separate image viewing activity.